requirements.txt
# Core deep learning
torch>=2.0,<2.5
torch-geometric>=2.4

# Datasets
ogb>=1.3.6                 # ogbn-arxiv
dgl>=1.1                   # YelpChi / Amazon (FraudYelp / FraudAmazon)
networkx>=2.8              # used by anomaly_injection (community detection fallback)
python-louvain>=0.16       # optional; community detection for "contextual" anomalies

# Numerics & evaluation
numpy>=1.23
scikit-learn>=1.2
scipy>=1.10

# Configuration & I/O
pyyaml>=6.0

# Visualization
matplotlib>=3.6
